Add optional rocket dependency to implement FromParam for World

+++ assets/pre-commit.ps1
@@ -6,6 +6,12 @@ if (-not $?)
     throw 'Native Failure'
 }
 
+cargo check --all-features
+if (-not $?)
+{
+    throw 'Native Failure'
+}
+
 # copy the tree to the WSL file system to improve compile times
 wsl rsync --delete -av /mnt/c/Users/fenhl/git/github.com/wurstmineberg/systemd-minecraft/stage/ /home/fenhl/wslgit/github.com/wurstmineberg/systemd-minecraft/ --exclude target
 if (-not $?)
